Make the page with all the quotes
