+++ json_handler.py
@@ -1,5 +1,4 @@
 import json
-from Quote import Quote
 
 fp = "static/quotes.json"
 
@@ -13,33 +12,4 @@ def GetJson():
 
 def SetJson(structure):
 	with open(fp, 'w') as file:
-		json.dump(file, structure)
-
-
-def AddQuote(quote):
-	new_quote = {
-		"Id": quote.Id,
-		"Quote": quote.Quote,
-		"Source": {
-			"Person": quote["Person"],
-			"Episode": quote["Episode"]
-		}
-	}
-
-	structure = GetJson()
-	structure["quotes"].append(new_quote)
-
-	SetJson(structure)
-
-
-def GetQuotes():
-	quote_list = []
-
-	for quote in GetJson()["quotes"]:
-		quote_list.append(Quote(
-			quote=quote["Quote"],
-			source=quote["Source"],
-			id=quote["Id"])
-		)
-
-	return quote_list
+		file.write(json.dumps(structure, indent=4))
