+++ templates/index.html
@@ -2,7 +2,5 @@
 {% block title %}The owl house quotes{% endblock %}
 {% block body %}
 
-    <h1>Welcome page</h1>
-    <p>*Insert random quote here*</p>
-
+<h1>The Owl House Quotes</h1>>
 {% endblock %}
